+++ python/mxnet/onnx/export_model.py
@@ -22,8 +22,8 @@
 import logging
 import numpy as np
 
-from ....base import string_types
-from .... import symbol
+from ..base import string_types
+from .. import symbol
 from .export_onnx import MXNetGraph
 from ._export_helper import load_module
